Calculate interest accumulated over a period of time
The program takes principal, interest rate, and years as parameters to
calculate the value using the formula principal * (1 + interest rate).
The program requires the interest rate to be a float data type and check
if all the variables are valid, greater than 0. The function iterates the
formula for the number of years and returns the amount incured and returns
false if the parameters are invalid.