Calculator program to process operation on two numbers
Calculator function takes number1, number2, and operator as parameters.
The function matches the operator passed and conducts the required
calculations over the two input numbers and exits the program right away
if the operator is not valid. It performs the following operations:

-If the operator is '+': it adds the two numbers

-If the operator is '-': it subtracts the two numbers

-If the operator is '*': it multiplies the two numbers

-If the operator is '/': it divides the two numbers

-If the operator is '//': it performs integer division

-If the operator is '**': it performs power operation

The input-output function allows the user to enter the two numbers and the
operator during runtime and passes the values to the calculator function.
The function also asks the user if they want to continue or terminate the
process and loops the process until user enter 'y' to terminate.