Decorator to time function execution
The calculate_time decorator takes a function as a parameter and uses
time.time() function from time module to calculate the start and end time
with the function executed inbetween. It subtracts the start time from the
end time to find total time taken for the execution of the passed function
and returns the wrapper funtion used to modify the passed function. To use
the decorator add '@calculate_time' before defining the function.